requirements.txt
pandas
chromadb
ollama
psycopg2-binary
python-dotenv
timescale-vector
instructor
anthropic
requests
pydantic
